Customize: Move UserModuleReloader to its own module
